% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_elemental_compositions.R
\name{calcNominalMass}
\alias{calcNominalMass}
\title{Predict the nominal mass}
\usage{
calcNominalMass(nbS, monoMass, alpha)
}
\arguments{
\item{nbS}{number of S-atoms}

\item{monoMass}{neutral monoisotopic mass of the protein or peptide of interest}

\item{alpha}{significance level, used to create 1-alpha prediction intervals}
}
\value{
A vector with the predicted nominal mass, its predicted lower limit and upper limit
}
\description{
Predict the nominal mass
}
\details{
The function is valid for proteins and peptides with a mass between 0 dalton and 4000 dalton.

The parameter alpha accepts 0.05 and 0.01 as values. In case another significance value is chosen, it is automatically converted to 0.05.
}
\examples{
calcNominalMass(nbS=0, monoMass=1709.897, alpha=0.05)
calcNominalMass(nbS=0, monoMass=1709.897, alpha=0.01)
calcNominalMass(nbS=0, monoMass=1709.897, alpha=0.15)
}
