% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacmass.R
\name{pacmass}
\alias{pacmass}
\title{Predict the elemental composition of proteins and peptides}
\usage{
pacmass(mz = NULL, z = NULL, fileNAME = NULL, nbS, ppm = 5, alpha = 0.05)
}
\arguments{
\item{mz}{monoisotopic mass-to-charge value}

\item{z}{charge}

\item{fileNAME}{name of a file containing mz and z.}

\item{nbS}{number of S-atoms}

\item{ppm}{mass error tolerance}

\item{alpha}{significance level}
}
\value{
A list with the predicted elemental compositions
}
\description{
Predict the elemental composition of proteins and peptides
}
\details{
The default value of mz, and fileNAME is NULL. However one of these parameters has to be defined in order for pacmass to work.

When mz is defined and z=NULL, it is assumed that mz is the mass of a neutral molecule.

When fileNAME is defined, the file has to be a comma-separated (csv) file, a tab-separated txt file or a tab-separated tsv file. The mass values have to be in a column with as name 'mz', the charges have to be in column with as name 'z'.
}
\examples{
pacmass(mz=1709.897, nbS=0, ppm=10, alpha=0.05)
pacmass(mz=855.956, z=2, nbS=0)
pacmass(mz=855.956, z=2, nbS=c(0,1))
pacmass(mz=c(855.956,735.693), z=c(2,3), nbS=0)
}
